[UPD] Split files for better location [ADD] gulp-concat to concatenate files & uglify
