+++ package.json
@@ -1,29 +1,30 @@
 {
-    "name": "angular-swiper",
-    "version": "0.4.0",
-    "description": "Angular directive for nolimits4web/swiper",
-    "main": "dist/angular-swiper.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/ksachdeva/angular-swiper.git"
-    },
-    "keywords": [
-        "angular",
-        "swiper"
-    ],
-    "author": "Kapil Sachdeva",
-    "license": "Apache",
-    "bugs": {
-        "url": "https://github.com/ksachdeva/angular-swiper/issues"
-    },
-    "homepage": "https://github.com/ksachdeva/angular-swiper",
-    "devDependencies": {
-        "gulp": "^3.8.11",
-        "gulp-ng-annotate": "^0.5.2",
-        "gulp-uglify": "^1.1.0",
-        "gulp-webserver": "^0.9.0"
-    }
+  "name": "angular-swiper",
+  "version": "0.4.0",
+  "description": "Angular directive for nolimits4web/swiper",
+  "main": "dist/angular-swiper.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ksachdeva/angular-swiper.git"
+  },
+  "keywords": [
+    "angular",
+    "swiper"
+  ],
+  "author": "Kapil Sachdeva",
+  "license": "Apache",
+  "bugs": {
+    "url": "https://github.com/ksachdeva/angular-swiper/issues"
+  },
+  "homepage": "https://github.com/ksachdeva/angular-swiper",
+  "devDependencies": {
+    "gulp": "^3.8.11",
+    "gulp-concat": "^2.6.0",
+    "gulp-ng-annotate": "^0.5.2",
+    "gulp-uglify": "^1.1.0",
+    "gulp-webserver": "^0.9.0"
+  }
 }
