+++ src/angular-swiper.module.js
@@ -1,0 +1,8 @@
+(function(window, angular, undefined) {
+
+    'use strict';
+
+    angular
+        .module('ksSwiper', []);
+
+})(window, angular, undefined);
