+++ src/angular-swiper.provider.js
@@ -1,0 +1,13 @@
+(function(window, angular, undefined) {
+
+    'use strict';
+
+    angular
+        .module('ksSwiper')
+        .provider('angularSwiperConfig', angularSwiperConfig);
+
+    function angularSwiperConfig() {
+
+    }
+
+})(window, angular, undefined);
